refactor: split organizeLibrary into focused modules
